Immutable (Read-Only Fields)
Make fields on `Record`s and `Meta` that are meant to be read-only _actually_ read-only by way of protected fields and magic getters. This also requires the use of a custom JSON serialization routine to ensure the fields flow into the appropriate places.

Also extend both serializable objects with proper deserialization routines to clean up all of the uses of `json_decode()` within the codebase.